Bake build identifiers into openwrt img
Added /etc/build.info and /etc/templates.info to reflect when each
step was run. This way the scale team will have the ability to track
when a particular img was build along with went the templates were
compile. Ideally both are done at the same time but its not necessarily
the case.